% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betterTaxaNames.R
\name{betterGeneraNames}
\alias{betterGeneraNames}
\title{More concise names for (greengenes format) microbial genera}
\usage{
betterGeneraNames(class, order, family, genus)
}
\arguments{
\item{order}{(vector of) order names}

\item{family}{(vector of) family names}

\item{genus}{(vector of) genus names}
}
\value{
simplified genus names
}
\description{
More concise names for (greengenes format) microbial genera
}
